Summary: Wrote Map class, Ant class, and some code for AntColony
Need to differentiate between elitist / regular ant functionalities

At the point just before calculating probabilities. What I am doing right now is I am implementing elitist ant system and running the runACO function in main because that is the only way we can pass information correctly. In the runACO function, I am getting the distance between the current city and the randomly selected city from cities not yet visited. Basically, if a city index shows up in the map of visitedCities, we don't take it. We go until we find a city that is not in the map.

Once we find a city not in the map, we get the distance between the cities, and call antColony.citiesSatisfyProbability to do our calculations in the AntColony class. Then if we get a number that lets us break out of the loop, we transfer control back to the antColony class.